Use model for SelectCustom state
Before this change, the SelectCustom component had an internalState variable,
that needed to be in sync with the value property provided by the
owning component.

This design turned out to be problematic, e.g. resetting the state by
the owning component was buggy.

This commit now uses defineModel from Vue for internal state to
solve the bug mentioned above and to simplify the implementation.